mayastor: add simple scheduling logic

+++ mayastor/src/bdev/nexus/nexus_fn_table.rs
@@ -165,7 +165,7 @@ impl NexusFnTable {
             spdk_json_write_val_raw(
                 w,
                 data.as_ptr() as *const _,
-                data.as_bytes().len(),
+                data.as_bytes().len() as u64,
             );
 
             spdk_json_write_array_end(w);
